requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
slack-bolt>=1.18.0
flask>=3.0.0
discord.py>=2.3.2 